Accidentally merge changes into the main tree.  Reverted.  Also fixed a few recent compile problems.
Former-commit-id: ccaae329495850009d68eecf0c3cfc460e213d9b

+++ WhirlyGlobeSrc/WhirlyGlobeLib/src/FontTextureManager.mm
@@ -389,7 +389,7 @@ typedef std::set<DrawStringRep *,IdentifiableSorter> DrawStringRepSet;
 
     // Work through the runs (which share attributes)
     CFArrayRef runs = CTLineGetGlyphRuns(line);
-    float lineHeight = 0.0, lineWidth = 0.0, ascent = 0.0, descent = 0.0;
+    CGFloat lineHeight = 0.0, lineWidth = 0.0, ascent = 0.0, descent = 0.0;
     lineWidth = CTLineGetTypographicBounds(line,&ascent,&descent,NULL);
     lineHeight = ascent+descent;
